Reworked how full network record are generated
Turned out s simpler way of generating a complete record every X times instead of trying to be clever was more optimal in terms of datastore writes.
A complete record is now generated once every hours.
Added a boolean in Network to state wether of not it represents a complete record.
Droppeda few persistent fields in StationProperties
Made locked and installed StationProperties non persistent as they are now encoded in bike availability