Genesis my ass. I couldn't setup CRON jobs the way I originally intended so I came up with a more flexible approach.
** this is the version that will be deployed on all instances with specific time setup**

Giving you :

Parsing status control

The whole parsing process can now  be activated by making calls to specific Cron URLs (requires administrator privileges)
As follow
http://localshost:8080/cron/activateparsingcronjob?activate=true
http://localshost:8080/cron/activateparsingcronjob?activate=false

By default, parsing_active is FALSE, meaning parsing job will be called but do nothing.
It is so because the cron.xml entries for the activation job are commented. To setup autocron in a DEPLOYED environment (no support from GAE for cron in local environment, just call the URLs yourself, or have your own cron doing it)

You can check the parsing_active current status via a request on the API (getstats, (and future isparsingactive)).

I added an entity to the datastore to track the status and log some data about the parsing process. It's all available through the API endpoint method getstats

Future refactoring : clean the mess I made between DataStats and ParsingStatus. I also could optimise the read cost of getstats by not having to pass queries to grab times of oldest/newest records. I leave it as is right now given time constraints and the fact getstats is not expected to be called that often.